Move updatable playbook types to /models
