+++ src/client.spec.ts
@@ -32,7 +32,6 @@ import {
 	SystemStatusCategory,
 	SystemStatusMessageReceived,
 	SystemStatusMessageSent,
-	UpdatablePlaybook,
 	UpdatableResource,
 	UpdatableSavedQuery,
 	UpdatableScheduledQuery,
@@ -88,6 +87,7 @@ import {
 	UpdatableGroup,
 	UpdatableMacro,
 	UpdatableNotification,
+	UpdatablePlaybook,
 	UploadableAutoExtractor,
 	User,
 	UserPreferences,
