+++ src/functions/playbooks/update-one-playbook.spec.ts
@@ -7,14 +7,14 @@
  **************************************************************************/
 
 import { omit } from 'lodash';
-import { CreatablePlaybook, isPlaybook, Playbook } from '../../models';
+import { CreatablePlaybook, isPlaybook, Playbook, UpdatablePlaybook } from '../../models';
 import { integrationTest } from '../../tests';
 import { TEST_AUTH_TOKEN, TEST_HOST } from '../../tests/config';
 import { omitUndefinedShallow } from '../utils';
 import { makeCreateOnePlaybook } from './create-one-playbook';
 import { makeDeleteOnePlaybook } from './delete-one-playbook';
 import { makeGetOnePlaybook } from './get-one-playbook';
-import { makeUpdateOnePlaybook, UpdatablePlaybook } from './update-one-playbook';
+import { makeUpdateOnePlaybook } from './update-one-playbook';
 
 describe('updateOnePlaybook()', () => {
 	const createOnePlaybook = makeCreateOnePlaybook({ host: TEST_HOST, useEncryption: false });
