+++ src/functions/playbooks/update-one-playbook.ts
@@ -6,17 +6,13 @@
  * MIT license. See the LICENSE file for details.
  **************************************************************************/
 
-import { encode as base64Encode } from 'base-64';
-import { isNull, isUndefined } from 'lodash';
-import { Playbook, RawPlaybook, RawPlaybookDecodedMetadata, toPlaybook } from '../../models';
-import { isUUID, Markdown, NumericID, RawNumericID, toRawNumericID, UUID } from '../../value-objects';
+import { Playbook, RawPlaybook, toPlaybook, toRawUpdatablePlaybook, UpdatablePlaybook } from '../../models';
 import {
 	APIFunctionMakerOptions,
 	buildHTTPRequest,
 	buildURL,
 	fetch,
 	HTTPRequestOptions,
-	omitUndefinedShallow,
 	parseJSONResponse,
 } from '../utils';
 import { makeGetOnePlaybook } from './get-one-playbook';
@@ -49,58 +45,5 @@ export const makeUpdateOnePlaybook = (makerOptions: APIFunctionMakerOptions) => 
 			if (err instanceof Error) throw err;
 			throw Error('Unknown error');
 		}
-	};
-};
-
-export interface UpdatablePlaybook {
-	uuid: UUID;
-
-	userID?: NumericID;
-	groupIDs?: Array<NumericID>;
-
-	name?: string;
-	description?: string | null;
-	labels?: Array<string>;
-
-	isGlobal?: boolean;
-
-	body?: Markdown;
-	coverImageFileGUID?: UUID | null;
-}
-
-interface RawUpdatablePlaybook {
-	Name: string;
-	Desc: string | null;
-	Body: string; // Base64 encoded markdown string
-	Metadata: string; // Base64 encoded RawPlaybookDecodedMetadata
-
-	UID: RawNumericID;
-	GIDs: Array<RawNumericID>;
-
-	Global: boolean;
-	Labels: Array<string>;
-}
-
-const toRawUpdatablePlaybook = (updatable: UpdatablePlaybook, current: Playbook): RawUpdatablePlaybook => {
-	const metadata: RawPlaybookDecodedMetadata = { dashboards: [] };
-	if (isUUID(current.coverImageFileGUID))
-		metadata.attachments = [{ context: 'cover', type: 'image', fileGUID: current.coverImageFileGUID }];
-
-	if (isNull(updatable.coverImageFileGUID)) metadata.attachments = undefined;
-	else if (isUUID(updatable.coverImageFileGUID))
-		metadata.attachments = [{ context: 'cover', type: 'image', fileGUID: updatable.coverImageFileGUID }];
-
-	return {
-		UID: toRawNumericID(updatable.userID ?? current.userID),
-		GIDs: (updatable.groupIDs ?? current.groupIDs).map(toRawNumericID),
-
-		Global: updatable.isGlobal ?? current.isGlobal,
-		Labels: updatable.labels ?? current.labels,
-
-		Name: updatable.name ?? current.name,
-		Desc: isUndefined(updatable.description) ? current.description : updatable.description,
-
-		Body: base64Encode(updatable.body ?? current.body),
-		Metadata: base64Encode(JSON.stringify(omitUndefinedShallow(metadata))),
 	};
 };
