+++ src/models/playbook/raw-updatable-playbook.ts
@@ -1,0 +1,22 @@
+/*************************************************************************
+ * Copyright 2020 Gravwell, Inc. All rights reserved.
+ * Contact: <legal@gravwell.io>
+ *
+ * This software may be modified and distributed under the terms of the
+ * MIT license. See the LICENSE file for details.
+ **************************************************************************/
+
+import { RawNumericID } from '../../value-objects';
+
+export interface RawUpdatablePlaybook {
+	Name: string;
+	Desc: string | null;
+	Body: string; // Base64 encoded markdown string
+	Metadata: string; // Base64 encoded RawPlaybookDecodedMetadata
+
+	UID: RawNumericID;
+	GIDs: Array<RawNumericID>;
+
+	Global: boolean;
+	Labels: Array<string>;
+}
